refactor(RemoveUsersWithoutEmail): delete users one by one

+++ lib/skate/remove_users_without_email.ex
@@ -8,11 +8,16 @@ defmodule Skate.RemoveUsersWithoutEmail do
   alias Skate.Settings.Db.User
 
   def run() do
-    Repo.delete_all(
-      from(user in User,
-        where: is_nil(user.email) or user.email == "",
-        select: user.id
+    user_ids_to_delete =
+      Repo.all(
+        from(user in User,
+          where: is_nil(user.email) or user.email == "",
+          select: user.id
+        )
       )
-    )
+
+    Enum.each(user_ids_to_delete, fn user_id ->
+      Repo.delete_all(from(user in User, where: user.id == ^user_id))
+    end)
   end
 end
